+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/backupAuton/auton/time_nearHuman.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auton;
+package org.firstinspires.ftc.teamcode.backupAuton.auton;
 //this is TIME BASED
 // TODO CHANGE NUMBERS AFTER
 import com.acmerobotics.dashboard.config.Config;
