--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rr/messages/PoseMessage.java
@@ -1,18 +1,0 @@
-package org.firstinspires.ftc.teamcode.rr.messages;
-
-import com.acmerobotics.roadrunner.Pose2d;
-
-public final class PoseMessage {
-    public long timestamp;
-    public double x;
-    public double y;
-    public double heading;
-
-    public PoseMessage(Pose2d pose) {
-        this.timestamp = System.nanoTime();
-        this.x = pose.position.x;
-        this.y = pose.position.y;
-        this.heading = pose.heading.toDouble();
-    }
-}
-
