--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rr/messages/TankCommandMessage.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.rr.messages;
-
-public final class TankCommandMessage {
-    public long timestamp;
-    public double voltage;
-    public double leftPower;
-    public double rightPower;
-
-    public TankCommandMessage(double voltage, double leftPower, double rightPower) {
-        this.timestamp = System.nanoTime();
-        this.voltage = voltage;
-        this.leftPower = leftPower;
-        this.rightPower = rightPower;
-    }
-}
